Remove interval-related aggregation data methods (#466)
addOther, multipleByFraction and clear were planned to be used
for the interval window implementation. Given the interval
window is removed, we should remove the aggregation methods
related to them.

Also switch MeanData.Count back to int64 which was converted
to float64 for fractions.

Fixes #465.

+++ zpages/rpcz.go
@@ -307,7 +307,7 @@ func (s snapExporter) ExportView(vd *view.Data) {
 			count = float64(v.Count)
 		case *view.MeanData:
 			sum = v.Sum()
-			count = v.Count
+			count = float64(v.Count)
 		case *view.SumData:
 			sum = float64(*v)
 			count = float64(*v)
